requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
phonenumbers==9.0.2
pillow==11.1.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
uritemplate==4.1.1
uuid==1.30
